refactor cell names
